+++ packages/ai-gateway/ai_gateway/providers/__init__.py
@@ -1,0 +1,9 @@
+from .ollama import OllamaProvider
+from .openai import OpenaiProvider
+from .jina import JinaProvider
+
+providers = {
+    "ollama": OllamaProvider(),
+    "openai": OpenaiProvider(),
+    "jina": JinaProvider(),
+}
